requirements.txt
Django==4.2.2
djangorestframework==3.14.0
python-decouple==3.8
Requests==2.31.0
